pre-commit running clean
