+++ src/saltext/consul/states/consul.py
@@ -103,7 +103,7 @@ def acl_present(
         "name": name,
         "changes": {},
         "result": True,
-        "comment": 'ACL "{}" exists and is up to date'.format(name),
+        "comment": f'ACL "{name}" exists and is up to date',
     }
 
     exists = _acl_exists(name, id, token, consul_url)
@@ -180,7 +180,7 @@ def acl_absent(name, id=None, token=None, consul_url="http://localhost:8500"):
         "name": id,
         "changes": {},
         "result": True,
-        "comment": 'ACL "{}" does not exist'.format(id),
+        "comment": f'ACL "{id}" does not exist',
     }
 
     exists = _acl_exists(name, id, token, consul_url)
@@ -190,9 +190,7 @@ def acl_absent(name, id=None, token=None, consul_url="http://localhost:8500"):
             ret["comment"] = "The acl exists, it will be deleted"
             return ret
 
-        delete = __salt__["consul.acl_delete"](
-            id=exists["id"], token=token, consul_url=consul_url
-        )
+        delete = __salt__["consul.acl_delete"](id=exists["id"], token=token, consul_url=consul_url)
         if delete["res"]:
             ret["result"] = True
             ret["comment"] = "The acl has been deleted"
