+++ tests/unit/states/test_consul.py
@@ -1,31 +1,159 @@
+"""
+Test case for the consul state module
+"""
+import logging
+from unittest.mock import MagicMock
+from unittest.mock import patch
+
 import pytest
-import salt.modules.test as testmod
-import saltext.consul.modules.consul_mod as consul_module
-import saltext.consul.states.consul_mod as consul_state
+import salt.states.consul as consul
+
+log = logging.getLogger(__name__)
 
 
 @pytest.fixture
 def configure_loader_modules():
     return {
-        consul_module: {
-            "__salt__": {
-                "test.echo": testmod.echo,
+        consul: {
+            "__opts__": {
+                "consul": {"url": "http://127.0.0.1", "token": "test_token"},
+                "test": False,
             },
-        },
-        consul_state: {
-            "__salt__": {
-                "consul.example_function": consul_module.example_function,
-            },
-        },
+            "__grains__": {"id": "test-minion"},
+        }
     }
 
 
-def test_replace_this_this_with_something_meaningful():
-    echo_str = "Echoed!"
-    expected = {
-        "name": echo_str,
-        "changes": {},
-        "result": True,
-        "comment": f"The 'consul.example_function' returned: '{echo_str}'",
+def test_acl_present():
+    """
+    Test salt.states.consul.acl_present function
+    """
+    acl_info = {
+        "data": [
+            {
+                "ID": "1d53dcd6-8f4f-431e-818a-9987996701a1",
+                "Name": "89530557-8f18-4e29-a2d6-5b2fc8bca713",
+                "Type": "client",
+                "Rules": "",
+                "CreateIndex": 419,
+                "ModifyIndex": 420,
+            }
+        ],
+        "res": True,
     }
-    assert consul_state.exampled(echo_str) == expected
+
+    acl_info_mock = MagicMock(return_value=acl_info)
+    with patch.dict(consul.__salt__, {"consul.acl_info": acl_info_mock}):
+        with patch.object(consul, "_acl_changes", MagicMock(return_value=False)):
+            consul_return = consul.acl_present(
+                "my_acl",
+                id="1d53dcd6-8f4f-431e-818a-9987996701a1",
+                token="89530557-8f18-4e29-a2d6-5b2fc8bca713",
+                type="client",
+                consul_url="http://localhost:8500",
+            )
+
+            _expected = {
+                "changes": {},
+                "comment": 'ACL "my_acl" exists and is up to date',
+                "name": "my_acl",
+                "result": True,
+            }
+            assert consul_return == _expected
+
+        acl_update_mock = MagicMock(
+            return_value={
+                "data": {"ID": "1D53DCD6-8F4F-431E-818A-9987996701A1"},
+                "res": True,
+            }
+        )
+        with patch.object(consul, "_acl_changes", MagicMock(return_value=True)):
+            with patch.dict(consul.__salt__, {"consul.acl_update": acl_update_mock}):
+                consul_return = consul.acl_present(
+                    "my_acl",
+                    id="1d53dcd6-8f4f-431e-818a-9987996701a1",
+                    token="89530557-8f18-4e29-a2d6-5b2fc8bca713",
+                    type="client",
+                    consul_url="http://localhost:8500",
+                )
+
+            _expected = {
+                "changes": {},
+                "comment": "The acl has been updated",
+                "name": "my_acl",
+                "result": True,
+            }
+            assert consul_return == _expected
+
+
+def test_acl_absent():
+    """
+    Test salt.states.consul.acl_absent function
+    """
+    #
+    # Test when the ACL does exist
+    #
+    acl_info = {
+        "data": [
+            {
+                "ID": "1d53dcd6-8f4f-431e-818a-9987996701a1",
+                "Name": "89530557-8f18-4e29-a2d6-5b2fc8bca713",
+                "Type": "client",
+                "Rules": "",
+                "CreateIndex": 419,
+                "ModifyIndex": 420,
+            }
+        ],
+        "res": True,
+    }
+    acl_info_mock = MagicMock(return_value=acl_info)
+    acl_delete_mock = MagicMock(
+        return_value={
+            "res": True,
+            "message": "ACL 38AC8470-4A83-4140-8DFD-F924CD32917F deleted.",
+        }
+    )
+    with patch.dict(consul.__salt__, {"consul.acl_info": acl_info_mock}):
+        with patch.dict(consul.__salt__, {"consul.acl_delete": acl_delete_mock}):
+            consul_return = consul.acl_absent(
+                "my_acl",
+                id="1d53dcd6-8f4f-431e-818a-9987996701a1",
+                token="89530557-8f18-4e29-a2d6-5b2fc8bca713",
+                consul_url="http://localhost:8500",
+            )
+
+            _expected = {
+                "changes": {},
+                "comment": "The acl has been deleted",
+                "name": "1d53dcd6-8f4f-431e-818a-9987996701a1",
+                "result": True,
+            }
+            assert consul_return == _expected
+
+    #
+    # Test when the ACL does not exist
+    #
+    acl_info = {"data": [], "res": True}
+    acl_info_mock = MagicMock(return_value=acl_info)
+    acl_delete_mock = MagicMock(
+        return_value={
+            "res": True,
+            "message": "ACL 38AC8470-4A83-4140-8DFD-F924CD32917F deleted.",
+        }
+    )
+    with patch.dict(consul.__salt__, {"consul.acl_info": acl_info_mock}):
+        with patch.dict(consul.__salt__, {"consul.acl_delete": acl_delete_mock}):
+            consul_return = consul.acl_absent(
+                "my_acl",
+                id="1d53dcd6-8f4f-431e-818a-9987996701a1",
+                token="89530557-8f18-4e29-a2d6-5b2fc8bca713",
+                consul_url="http://localhost:8500",
+            )
+
+            _expected = {
+                "changes": {},
+                "comment": 'ACL "1d53dcd6-8f4f-431e-818a-9987996701a1" does not exist',
+                "name": "1d53dcd6-8f4f-431e-818a-9987996701a1",
+                "result": True,
+            }
+            assert consul_return == _expected
